move live777-core => liveion as lib
